requirements.txt
click
RPi.GPIO
hx711
